requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_pytorch_select=0.1=cpu_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
brotli=1.0.9=he6710b0_2
ca-certificates=2021.7.5=h06a4308_1
certifi=2021.5.30=py37h06a4308_0
cffi=1.14.6=py37h400218f_0
cycler=0.10.0=py37_0
dbus=1.13.18=hb2f20db_0
decorator=5.0.9=pyhd3eb1b0_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
glib=2.69.0=h5202010_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
intel-openmp=2019.4=243
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.26.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.0=py37h06a4308_1
jpeg=9b=h024ee3a_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py37h06a4308_0
kiwisolver=1.3.1=py37h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libmklml=2019.0.5=0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.4.2=py37h06a4308_0
matplotlib-base=3.4.2=py37hab158f2_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
munkres=1.1.4=py_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=hff7bd54_1
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py37_0
opencv-python=4.5.3.56=pypi_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py37h2c7a002_0
pip=21.1.3=py37h06a4308_0
prompt-toolkit=3.0.17=pyh06a4308_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.20=py_2
pygments=2.9.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.7.1=cpu_py37h6a09485_0
pyzmq=22.2.1=py37h295c915_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tk=8.6.10=hbc83047_0
torchvision=0.8.2=cpu_py37ha229d99_0
tornado=6.1=py37h27cfd23_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
wcwidth=0.2.5=py_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
